Add comprehensive tests for flight and user services, repositories, and routers
- Implemented unit tests for FlightBusinessService, covering flight search, creation, and listing functionalities.
- Created tests for UserRepository using an in-memory SQLite database to ensure data isolation.
- Developed tests for UserService, focusing on registration and login processes, including edge cases.
- Added tests for User Router to validate HTTP interactions and error handling.
- Ensured all tests mock dependencies to isolate business logic and verify expected behaviors.

+++ api/src/schemas/flight.py
@@ -24,7 +24,7 @@ class FlightResponse(BaseModel):
     arrival_time: datetime.datetime
     airline: str
     status: str
-    price: float
+    price: int  # Changed from float to int to match database model
 
     class Config:
         from_attributes = True
